combine composer errs files
